iot_greenhouse_codebase/requirements.txt
fastapi
uvicorn[standard]
numpy
joblib
pygam
influxdb-client
pydantic
azure-eventhub
azure-iot-device
